requirements.txt
fastf1==3.7.0
requests==2.32.5
urllib3==2.6.3
psycopg2-binary==2.9.11
